vision_model_playground/cnn_deployment/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
filelock==3.20.0
Flask==3.1.2
fsspec==2025.10.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
pillow==12.0.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
typing_extensions==4.15.0
Werkzeug==3.1.3
